arrays
array

+++ c programs/1D-AVG.c
@@ -1,22 +1,22 @@
 #include <stdio.h>
 int main()
 {
-     int marks[10], i, n, sum = 0, average;
+     float marks[10], i, n, sum = 0, average;
 
      printf("Enter number of elements: ");
-     scanf("%d", &n);
+     scanf("%f", &n);
 
      for(i=0; i<n; ++i)
      {
           printf("Enter number%d: ",i+1);
-          scanf("%d", &marks[i]);
+          scanf("%f", &marks[i]);
           
           // adding integers entered by the user to the sum variable
           sum += marks[i];
      }
 
      average = sum/n;
-     printf("Average = %d", average);
+     printf("Average = %f", average);
 
      return 0;
 }
